requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
awsebcli==3.21.0
blessed==1.20.0
botocore==1.35.36
cement==2.10.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
 
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
Django==5.1.1
django-admin-sortable2==2.2.8
django-markdown-deux==1.0.6
filelock==3.16.1
 
 
h11==0.14.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.3.0
markdown-it-py==3.0.0
markdown2==2.5.1
mdurl==0.1.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
parsel==1.9.1
pathspec==0.10.1
platformdirs==4.3.6
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
rich==13.8.1
 
 
gunicorn==23.0.0
whitenoise==6.7.0
semantic-version==2.10.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
termcolor==2.5.0
tldextract==5.1.2
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==1.26.20
w3lib==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
 
wsproto==1.2.0
yapf==0.40.2
zipp==3.20.2
zope.interface==7.0.3
